refactor(config): split load.ts into focused modules (#184)
load.ts mixed env-ref resolution, source detection, env overrides, YAML
parsing, .env fallback and a lightweight BLE loader in 419 lines. Split
each concern into its own module; load.ts becomes a 53-line thin
orchestration entrypoint that re-exports the public API, so every caller
and the existing tests keep importing from ./config/load.js unchanged.

- paths.ts          ROOT / DEFAULT_CONFIG_PATH / DEFAULT_ENV_PATH
- env-refs.ts        resolveEnvReferences (${VAR} expansion)
- source-detect.ts   ConfigSource + detectConfigSource
- env-overrides.ts   parseBleAdapterEnv / applyEnvOverrides / filterValidExporters
- yaml-load.ts       loadYamlConfig
- env-load.ts        loadEnvConfig (.env fallback)
- ble-load.ts        loadBleConfig + BleLoadedConfig
- load.ts            LoadedConfig + loadAppConfig + re-export barrel

Pure code move, no behaviour change: resolution order (YAML over env
over defaults) and dotenv's no-override default are preserved. Type-only
re-exports use `export type` (isolatedModules-safe). All 228 config
tests still green via the barrel + fs mock.
