+++ src/config/yaml-load.ts
@@ -1,0 +1,65 @@
+import { existsSync, readFileSync } from 'node:fs';
+import { parse as parseYaml } from 'yaml';
+import { config as dotenvConfig } from 'dotenv';
+import { createLogger } from '../logger.js';
+import { AppConfigSchema, formatConfigError } from './schema.js';
+import type { AppConfig } from './schema.js';
+import { DEFAULT_CONFIG_PATH, DEFAULT_ENV_PATH } from './paths.js';
+import { resolveEnvReferences } from './env-refs.js';
+import { applyEnvOverrides, filterValidExporters } from './env-overrides.js';
+
+const log = createLogger('Config');
+
+/**
+ * Load and validate config from a YAML file.
+ */
+export function loadYamlConfig(configPath?: string): AppConfig {
+  // Load .env so ${VAR} references in config.yaml can resolve secrets from .env
+  if (existsSync(DEFAULT_ENV_PATH)) {
+    dotenvConfig({ path: DEFAULT_ENV_PATH });
+  }
+
+  const yamlPath = configPath ?? DEFAULT_CONFIG_PATH;
+  const raw = readFileSync(yamlPath, 'utf8');
+  const parsed: unknown = parseYaml(raw);
+  const resolved = resolveEnvReferences(parsed);
+
+  const result = AppConfigSchema.safeParse(resolved);
+  if (!result.success) {
+    const msg = formatConfigError(result.error);
+    log.error(msg);
+    throw new Error(msg);
+  }
+
+  let config = result.data;
+
+  // Lenient exporter validation — warn + skip unknown types
+  config = {
+    ...config,
+    global_exporters: filterValidExporters(config.global_exporters),
+    users: config.users.map((u) => ({
+      ...u,
+      exporters: filterValidExporters(u.exporters),
+    })),
+  };
+
+  // Set NOBLE_DRIVER env var if configured (needed before BLE handler import)
+  if (config.ble?.noble_driver) {
+    process.env.NOBLE_DRIVER = config.ble.noble_driver;
+  }
+
+  // Set BLE_HANDLER env var if configured (needed before BLE handler import)
+  if (config.ble?.handler && config.ble.handler !== 'auto') {
+    process.env.BLE_HANDLER = config.ble.handler;
+  }
+
+  // Set DEBUG env var if configured (needed for logger level)
+  if (config.runtime?.debug) {
+    process.env.DEBUG = 'true';
+  }
+
+  // Apply env overrides
+  config = applyEnvOverrides(config);
+
+  return config;
+}
